requirements.txt
groq==0.8.0
python-dotenv==1.0.1
openai=1.31.1
